xml/parsing/manakai/notations-perefs-1.dat

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % fuga "xy">
<!NOTATION %fuga;SYSTEM "aa">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
#document
| <!DOCTYPE a "" "http://a/">
|   <!NOTATION xy "" "aa">
| <{} a>

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % fuga "xy">
<!NOTATION %fuga; SYSTEM "aa">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
#document
| <!DOCTYPE a "" "http://a/">
|   <!NOTATION xy "" "aa">
| <{} a>

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % fuga "xy  SYSTEM">
<!NOTATION %fuga;  "aa">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
#document
| <!DOCTYPE a "" "http://a/">
|   <!NOTATION xy "" "aa">
| <{} a>

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % bar ">">
<!ENTITY % fuga "&#x25;bar;">
<!NOTATION %fuga; SYSTEM "aa">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;17;mdc in pe in md;;;m
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % bar "">
<!NOTATION a SYSTEM "aa"%bar;>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
#document
| <!DOCTYPE a "" "http://a/">
|   <!NOTATION a "" "aa">
| <{} a>
